Remove vLLM rollout, move trainer/losses/utils/benchmarks to DeepSpeedExamples
