--- deepspeed/runtime/rlhf/trainer/__init__.py
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-# DeepSpeed Team
-"""deepspeed.runtime.rlhf.trainer — RLHF training-loop implementations."""
-
-from deepspeed.runtime.rlhf.trainer.base import RLHFTrainer  # noqa: F401
-from deepspeed.runtime.rlhf.trainer.opsd import OPSDTrainer  # noqa: F401
