--- deepspeed/runtime/rollout/_vllm_compat/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-# DeepSpeed Team
